+++ src/console/src/Input/InputDefinition.php
@@ -4,9 +4,15 @@ namespace Swoft\Console\Input;
 
 /**
  * Class InputDefinition
- * @since 2.0 refer inhere/console
+ * @since 2.0 refer inhere/console and symfony/console
  */
 class InputDefinition
 {
-
+    /**
+     * @return InputDefinition
+     */
+    public static function create(): self
+    {
+        return new self();
+    }
 }