+++ src/stdlib/src/Arrayable.php
@@ -12,5 +12,5 @@ interface Arrayable
      *
      * @return array
      */
-    public function toArray();
+    public function toArray(): array;
 }